REFACTOR and FIX: Sidebar, Window components
	- Sidebar components children now has its own file components.
	- Was added context fot Sidebar data and states.
	- All file components was renamed to lower case, to avoid import
	  errors from next.
